+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## [0.0.2] - 31 Aug 2019
 
-* VerticalMenuForAndroidTV and HyperLinkMenuItemWidget supported,
-example updated
+* VerticalMenuForAndroidTV and HyperLinkMenuItem supported,
+
+* Container parameters are passed to VerticalMenuForAndroidTV which are used to display
+the MenuItem
+
+* Example updated
 
 ## [0.0.1] - 31 Aug 2019